05.03/ЛВА1/requirements.txt
numpy==2.1.1
opencv-python==4.11.0.86
websocket-client==1.8.0
websockets==12.0
